sequelize mock 제거, test seed 생성

+++ server/src/loaders/sequelize.js
@@ -1,7 +1,10 @@
 const sequelize = require('@models');
 
 const sequelizeLoader = async () => {
-  // const useInit = process.env.INIT_DB === 'use';
+  if (process.env.NODE_ENV === 'test') {
+    return;
+  }
+
   try {
     await sequelize.sync({ force: true });
   } catch (e) {
